+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # CHANGELOG Edit Flow for Custom Bulk/Quick Edit
 
 ## master
-* TBD
+
+## 1.0.0
+* Add screenshot 1. Edit Flow field settings
+* Add screenshot 2. Quick Edit of Edit Flow attributes
+* Add screenshot 3. Bulk Edit of Edit Flow attributes
+* Load Edit Flow customizations from Custom Bulk/Quick Edit Premium
 
 ## 0.0.1
 * Initial code release (with WordPress SEO basis, needs Edit Flow)